requirement.txt
asgiref==3.6.0
attrs==22.2.0
black==23.3.0
click==8.1.3
Django==4.2
djangorestframework==3.14.0
drf-spectacular==0.26.1
inflection==0.5.1
jsonschema==4.17.3
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
